Crate xcm

Cross-Consensus Message format data structures.

Modules

latest

opaque

prelude

v0

Version 0 of the Cross-Consensus Message format data structures.

v1

XCM Version 1

v2

XCM Version 2

Structs

AlwaysV0

WrapVersion implementation which attempts to always convert the XCM to version 0 before wrapping it.

AlwaysV1

WrapVersion implementation which attempts to always convert the XCM to version 1 before wrapping it.

AlwaysV2

WrapVersion implementation which attempts to always convert the XCM to version 2 before wrapping it.

DoubleEncoded

Wrapper around the encoded and decoded versions of a value. Caches the decoded value once computed.

Enums

Unsupported

VersionedMultiAsset

A single MultiAsset value, together with its version code.

VersionedMultiAssets

A single MultiAssets value, together with its version code.

VersionedMultiLocation

A single MultiLocation value, together with its version code.

VersionedResponse

A single Response value, together with its version code.

VersionedXcm

A single XCM message, together with its version code.

Constants

MAX_XCM_DECODE_DEPTH

Maximum nesting level for XCM decoding.

Traits

GetWeight

IntoVersion

Attempt to convert self into a particular version of itself.

WrapVersion

Convert an Xcm datum into a VersionedXcm, based on a destination MultiLocation which will interpret it.

Type Definitions

AlwaysLatest

WrapVersion implementation which attempts to always convert the XCM to the latest version before wrapping it.

AlwaysRelease

WrapVersion implementation which attempts to always convert the XCM to the release version before wrapping it.

Version

A version of XCM.