Trait xcm::IntoVersion

pub trait IntoVersion: Sized {
    // Required method
    fn into_version(self, version: Version) -> Result<Self, ()>;

    // Provided method
    fn into_latest(self) -> Result<Self, ()> { ... }
}

Attempt to convert self into a particular version of itself.

Required Methods

fn into_version(self, version: Version) -> Result<Self, ()>

Consume self and return same value expressed in some particular version of XCM.

Provided Methods

fn into_latest(self) -> Result<Self, ()>

Consume self and return same value expressed the latest version of XCM.