Module xcm::v0

Version 0 of the Cross-Consensus Message format data structures.

Re-exports

pub use multi_location::MultiLocation::*;

Modules

opaque

prelude

A prelude for importing all types typically used when interacting with XCM messages.

Enums

AssetInstance

A general identifier for an instance of a non-fungible asset class.

BodyId

An identifier of a pluralistic body.

BodyPart

A part of a pluralistic body.

Error

Junction

A single item in a path to describe the relative location of a consensus system.

MultiAsset

A single general identifier for an asset.

MultiLocation

A relative path between state-bearing consensus systems.

NetworkId

A global identifier of an account-bearing consensus system.

Order

An instruction to be executed on some or all of the assets in holding, used by asset-related XCM messages.

OriginKind

Basically just the XCM (more general) version of ParachainDispatchOrigin.

Outcome

Outcome of an XCM execution.

Response

Response data to a query.

Xcm

Cross-Consensus Message: A message from one consensus system to another.

Traits

ExecuteXcm

Type of XCM message executor.

SendXcm

Utility for sending an XCM message.

Type Definitions

Result