Module xcm::v0::prelude

A prelude for importing all types typically used when interacting with XCM messages.

Re-exports

pub use super::OriginKind;

pub use super::Xcm;

pub use super::junction::Junction::*;

pub use super::multi_asset::AssetInstance::*;

pub use super::multi_asset::MultiAsset::*;

pub use super::multi_location::MultiLocation::*;

pub use super::order::Order::*;

pub use super::Junction::*;

pub use super::Xcm::*;

Enums

AssetInstance

A general identifier for an instance of a non-fungible asset class.

BodyId

An identifier of a pluralistic body.

MultiAsset

A single general identifier for an asset.

MultiLocation

A relative path between state-bearing consensus systems.

Order

An instruction to be executed on some or all of the assets in holding, used by asset-related XCM messages.

Outcome

Outcome of an XCM execution.

XcmError

Traits

ExecuteXcm

Type of XCM message executor.

SendXcm

Utility for sending an XCM message.

Type Definitions

XcmResult