Module xcm::v1::multiasset

Cross-Consensus Message format asset data structures.

This encompasses four types for representing assets:

MultiAsset: A description of a single asset, either an instance of a non-fungible or some amount of a fungible.

MultiAssets: A collection of MultiAssets. These are stored in a Vec and sorted with fungibles first.

Wild: A single asset wildcard, this can either be “all” assets, or all assets of a specific kind.

MultiAssetFilter: A combination of Wild and MultiAssets designed for efficiently filtering an XCM holding account.

Structs

MultiAsset

MultiAssets
A Vec of MultiAssets. There may be no duplicate fungible items in here and when decoding, they must be sorted.

Enums

AssetId
Classification of an asset being concrete or abstract.

AssetInstance
A general identifier for an instance of a non-fungible asset class.

Fungibility
Classification of whether an asset is fungible or not, along with a mandatory amount or instance.

MultiAssetFilter
MultiAsset collection, either MultiAssets or a single wildcard.

WildFungibility
Classification of whether an asset is fungible or not.

WildMultiAsset
A wildcard representing a set of assets.