Enum xcm::v1::prelude::XcmError

pub enum XcmError {

    Undefined,
    Overflow,
    Unimplemented,
    UnhandledXcmVersion,
    UnhandledXcmMessage,
    UnhandledEffect,
    EscalationOfPrivilege,
    UntrustedReserveLocation,
    UntrustedTeleportLocation,
    DestinationBufferOverflow,
    SendFailed(&'static str),
    CannotReachDestination(MultiLocation, Xcm<()>),
    MultiLocationFull,
    FailedToDecode,
    BadOrigin,
    ExceedsMaxMessageSize,
    FailedToTransactAsset(&'static str),
    WeightLimitReached(u64),
    Wildcard,
    MaxWeightInvalid,
    NotHoldingFees,
    WeightNotComputable,
    Barrier,
    NotWithdrawable,
    LocationCannotHold,
    TooExpensive,
    AssetNotFound,
    DestinationUnsupported,
    RecursionLimitReached,

}

Variants

Undefined

Overflow

An arithmetic overflow happened.

Unimplemented

The operation is intentionally unsupported.

UnhandledXcmVersion

UnhandledXcmMessage

The implementation does not handle a given XCM.

UnhandledEffect

The implementation does not handle an effect present in an XCM.

EscalationOfPrivilege

UntrustedReserveLocation

UntrustedTeleportLocation

DestinationBufferOverflow

SendFailed(&'static str)

The message and destination was recognized as being reachable but the operation could not be completed. A human-readable explanation of the specific issue is provided.

CannotReachDestination(MultiLocation, Xcm<()>)

The message and destination combination was not recognized as being reachable.

MultiLocationFull

FailedToDecode

BadOrigin

ExceedsMaxMessageSize

FailedToTransactAsset(&'static str)

An asset transaction (like withdraw or deposit) failed. See implementers of the TransactAsset trait for sources. Causes can include type conversion failures between id or balance types.

WeightLimitReached(u64)

Execution of the XCM would potentially result in a greater weight used than the pre-specified weight limit. The amount that is potentially required is the parameter.

Wildcard

An asset wildcard was passed where it was not expected (e.g. as the asset to withdraw in a WithdrawAsset XCM).

MaxWeightInvalid

The case where an XCM message has specified a weight limit on an interior call and this limit is too low.

Used by:

Transact

NotHoldingFees

The fees specified by the XCM message were not found in the holding register.

Used by:

BuyExecution

WeightNotComputable

The weight of an XCM message is not computable ahead of execution. This generally means at least part of the message is invalid, which could be due to it containing overly nested structures or an invalid nested data segment (e.g. for the call in Transact).

Barrier

The XCM did not pass the barrier condition for execution. The barrier condition differs on different chains and in different circumstances, but generally it means that the conditions surrounding the message were not such that the chain considers the message worth spending time executing. Since most chains lift the barrier to execution on appropriate payment, presentation of an NFT voucher, or based on the message origin, it means that none of those were the case.

NotWithdrawable

Indicates that it is not possible for a location to have an asset be withdrawn or transferred from its ownership. This probably means it doesn’t own (enough of) it, but may also indicate that it is under a lock, hold, freeze or is otherwise unavailable.

LocationCannotHold

Indicates that the consensus system cannot deposit an asset under the ownership of a particular location.

TooExpensive

The assets given to purchase weight is are insufficient for the weight desired.

AssetNotFound

The given asset is not handled.

DestinationUnsupported

The given message cannot be translated into a format that the destination can be expected to interpret.

RecursionLimitReached

execute_xcm has been called too many times recursively.