Trait xcm::v1::ExecuteXcm

pub trait ExecuteXcm<Call> {
    fn execute_xcm_in_credit(
        origin: impl Into<MultiLocation>, 
        message: Xcm<Call>, 
        weight_limit: u64, 
        weight_credit: u64
    ) -> Outcome;

    fn execute_xcm(
        origin: impl Into<MultiLocation>, 
        message: Xcm<Call>, 
        weight_limit: u64
    ) -> Outcome { ... }
}

Type of XCM message executor.

Required methods

fn execute_xcm_in_credit(
origin: impl Into<MultiLocation>,
message: Xcm<Call>,
weight_limit: u64,
weight_credit: u64
) -> Outcome

Execute some XCM message from origin using no more than weight_limit weight.

Some amount of weight_credit may be provided which, depending on the implementation, may allow execution without associated payment.

Provided methods

fn execute_xcm(
origin: impl Into<MultiLocation>,
message: Xcm<Call>,
weight_limit: u64
) -> Outcome

Execute some XCM message from origin using no more than weight_limit weight. The weight limit is a basic hard-limit and the implementation may place further restrictions or requirements on weight and other aspects.