Enum xcm::v0::MultiLocation

pub enum MultiLocation {
    Null,
    X1(Junction),
    X2(Junction, Junction),
    X3(Junction, Junction, Junction),
    X4(Junction, Junction, Junction, Junction),
    X5(Junction, Junction, Junction, Junction, Junction),
    X6(Junction, Junction, Junction, Junction, Junction, Junction),
    X7(Junction, Junction, Junction, Junction, Junction, Junction, Junction),
    X8(Junction, Junction, Junction, Junction, Junction, Junction, Junction, Junction),
}

A relative path between state-bearing consensus systems.

A location in a consensus system is defined as an isolatable state machine held within global consensus. The location in question need not have a sophisticated consensus algorithm of its own; a single account within Ethereum, for example, could be considered a location.

A very-much non-exhaustive list of types of location include:

A (normal, layer-1) block chain, e.g. the Bitcoin mainnet or a parachain.

A layer-0 super-chain, e.g. the Polkadot Relay chain.

A layer-2 smart contract, e.g. an ERC-20 on Ethereum.

A logical functional component of a chain, e.g. a single instance of a pallet on a Frame-based Substrate chain.

An account.

A MultiLocation is a relative identifier, meaning that it can only be used to define the relative path between two locations, and cannot generally be used to refer to a location universally. It is comprised of a number of junctions, each morphing the previous location, either diving down into one of its internal locations, called a sub-consensus, or going up into its parent location. Correct MultiLocation values must have all Parent junctions as a prefix to all sub-consensus junctions.

This specific MultiLocation implementation uses a Rust enum in order to make pattern matching easier.

The MultiLocation value of Null simply refers to the interpreting consensus system.

Variants

Null

The interpreting consensus system.

X1(Junction)

A relative path comprising 1 junction.

X2(Junction, Junction)

A relative path comprising 2 junctions.

X3(Junction, Junction, Junction)

A relative path comprising 3 junctions.

X4(Junction, Junction, Junction, Junction)

A relative path comprising 4 junctions.

X5(Junction, Junction, Junction, Junction, Junction)

A relative path comprising 5 junctions.

X6(Junction, Junction, Junction, Junction, Junction, Junction)

A relative path comprising 6 junctions.

X7(Junction, Junction, Junction, Junction, Junction, Junction, Junction)

A relative path comprising 7 junctions.

X8(Junction, Junction, Junction, Junction, Junction, Junction, Junction, Junction)

A relative path comprising 8 junctions.

Implementations

impl MultiLocation

pub fn first(&self) -> Option<&Junction>

Returns first junction, or None if the location is empty.

pub fn last(&self) -> Option<&Junction>

Returns last junction, or None if the location is empty.

pub fn split_first(self) -> (MultiLocation, Option<Junction>)

Splits off the first junction, returning the remaining suffix (first item in tuple) and the first element (second item in tuple) or None if it was empty.

pub fn split_last(self) -> (MultiLocation, Option<Junction>)

Splits off the last junction, returning the remaining prefix (first item in tuple) and the last element (second item in tuple) or None if it was empty.

pub fn take_first(&mut self) -> Option<Junction>

Removes the first element from self, returning it (or None if it was empty).

pub fn take_last(&mut self) -> Option<Junction>

Removes the last element from self, returning it (or None if it was empty).

pub fn pushed_with(self, new: Junction) -> Result<Self, Self>

Consumes self and returns a MultiLocation suffixed with new, or an Err with the original value of self in case of overflow.

pub fn pushed_front_with(self, new: Junction) -> Result<Self, Self>

Consumes self and returns a MultiLocation prefixed with new, or an Err with the original value of self in case of overflow.

pub fn len(&self) -> usize

Returns the number of junctions in self.

pub fn at(&self, i: usize) -> Option<&Junction>

Returns the junction at index i, or None if the location doesn’t contain that many elements.

pub fn at_mut(&mut self, i: usize) -> Option<&mut Junction>

Returns a mutable reference to the junction at index i, or None if the location doesn’t contain that many elements.

pub fn iter(&self) -> MultiLocationRefIterator<'_>

Returns a reference iterator over the junctions.

pub fn iter_rev(&self) -> MultiLocationReverseRefIterator<'_>

Returns a reference iterator over the junctions in reverse.

pub fn into_iter(self) -> MultiLocationIterator

Consumes self and returns an iterator over the junctions.

pub fn into_iter_rev(self) -> MultiLocationReverseIterator

Consumes self and returns an iterator over the junctions in reverse.

pub fn match_and_split(&self, prefix: &MultiLocation) -> Option<&Junction>

Ensures that self begins with prefix and that it has a single Junction item following. If so, returns a reference to this Junction item.

Example

let mut m = X3(Parent, PalletInstance(3), OnlyChild);
assert_eq!(m.match_and_split(&X2(Parent, PalletInstance(3))), Some(&OnlyChild));
assert_eq!(m.match_and_split(&X1(Parent)), None);

pub fn push(&mut self, new: Junction) -> Result<(), ()>

Mutates self, suffixing it with new. Returns Err in case of overflow.

pub fn push_front(&mut self, new: Junction) -> Result<(), ()>

Mutates self, prefixing it with new. Returns Err in case of overflow.

pub fn leading_parent_count(&self) -> usize

Returns the number of Parent junctions at the beginning of self.

pub fn canonicalize(&mut self)

This function ensures a multi-junction is in its canonicalized/normalized form, removing any internal [Non-Parent, Parent] combinations.

pub fn append_with(&mut self, suffix: MultiLocation) -> Result<(), MultiLocation>

Mutate self so that it is suffixed with suffix. The correct normalized form is returned, removing any internal [Non-Parent, Parent] combinations.

In the case of overflow, self is unmodified and we return Err with suffix.

Example

let mut m = X3(Parent, Parachain(21), OnlyChild);
assert_eq!(m.append_with(X2(Parent, PalletInstance(3))), Ok(()));
assert_eq!(m, X3(Parent, Parachain(21), PalletInstance(3)));

pub fn prepend_with(
&mut self,
prefix: MultiLocation
) -> Result<(), MultiLocation>

Mutate self so that it is prefixed with prefix. The correct normalized form is returned, removing any internal [Non-Parent, Parent] combinations.

In the case of overflow, self is unmodified and we return Err with prefix.

Example

let mut m = X3(Parent, Parent, PalletInstance(3));
assert_eq!(m.prepend_with(X3(Parent, Parachain(21), OnlyChild)), Ok(()));
assert_eq!(m, X2(Parent, PalletInstance(3)));

pub fn is_interior(&self) -> bool

Returns true iff self is an interior location. For this it may not contain any Junctions for which Junction::is_interior returns false. This is generally true, except for the Parent item.

Example

let parent = X1(Parent);
assert_eq!(parent.is_interior(), false);
let m = X2(PalletInstance(12), AccountIndex64 { network: Any, index: 23 });
assert_eq!(m.is_interior(), true);