Enum xcm::v1::Xcm

pub enum Xcm<Call> {

    WithdrawAsset {
        assets: MultiAssets,
        effects: Vec<Order<Call>>,
    },
    ReserveAssetDeposited {
        assets: MultiAssets,
        effects: Vec<Order<Call>>,
    },
    ReceiveTeleportedAsset {
        assets: MultiAssets,
        effects: Vec<Order<Call>>,
    },
    QueryResponse {
        query_id: u64,
        response: Response,
    },
    TransferAsset {
        assets: MultiAssets,
        beneficiary: MultiLocation,
    },
    TransferReserveAsset {
        assets: MultiAssets,
        dest: MultiLocation,
        effects: Vec<Order<()>>,
    },
    Transact {
        origin_type: OriginKind,
        require_weight_at_most: u64,
        call: DoubleEncoded<Call>,
    },
    HrmpNewChannelOpenRequest {
        sender: u32,
        max_message_size: u32,
        max_capacity: u32,
    },
    HrmpChannelAccepted {
        recipient: u32,
    },
    HrmpChannelClosing {
        initiator: u32,
        sender: u32,
        recipient: u32,
    },
    RelayedFrom {
        who: InteriorMultiLocation,
        message: Box<Xcm<Call>>,
    },
    SubscribeVersion {
        query_id: u64,
        max_response_weight: u64,
    },
    UnsubscribeVersion,

}

Cross-Consensus Message: A message from one consensus system to another.

Consensus systems that may send and receive messages include blockchains and smart contracts.

All messages are delivered from a known origin, expressed as a MultiLocation.

This is the inner XCM format and is version-sensitive. Messages are typically passed using the outer XCM format, known as VersionedXcm.

Variants

WithdrawAsset

Withdraw asset(s) (assets) from the ownership of origin and place them into holding. Execute the orders (effects).

assets: The asset(s) to be withdrawn into holding.

effects: The order(s) to execute on the holding register.

Kind: Instruction.

Errors:

Fields of WithdrawAsset

assets: MultiAssetseffects: Vec<Order<Call>>

ReserveAssetDeposited

Asset(s) (assets) have been received into the ownership of this system on the origin system.

Some orders are given (effects) which should be executed once the corresponding derivative assets have been placed into holding.

assets: The asset(s) that are minted into holding.

effects: The order(s) to execute on the holding register.

Safety: origin must be trusted to have received and be storing assets such that they may later be withdrawn should this system send a corresponding message.

Kind: Trusted Indication.

Errors:

Fields of ReserveAssetDeposited

assets: MultiAssetseffects: Vec<Order<Call>>

ReceiveTeleportedAsset

Asset(s) (assets) have been destroyed on the origin system and equivalent assets should be created on this system.

Some orders are given (effects) which should be executed once the corresponding derivative assets have been placed into the Holding Register.

assets: The asset(s) that are minted into the Holding Register.

effects: The order(s) to execute on the Holding Register.

Safety: origin must be trusted to have irrevocably destroyed the corresponding assets prior as a consequence of sending this message.

Kind: Trusted Indication.

Errors:

Fields of ReceiveTeleportedAsset

assets: MultiAssetseffects: Vec<Order<Call>>

QueryResponse

Indication of the contents of the holding register corresponding to the QueryHolding order of query_id.

query_id: The identifier of the query that resulted in this message being sent.

assets: The message content.

Safety: No concerns.

Kind: Information.

Errors:

Fields of QueryResponse

query_id: u64response: Response

TransferAsset

Withdraw asset(s) (assets) from the ownership of origin and place equivalent assets under the ownership of beneficiary.

assets: The asset(s) to be withdrawn.

beneficiary: The new owner for the assets.

Safety: No concerns.

Kind: Instruction.

Errors:

Fields of TransferAsset

assets: MultiAssetsbeneficiary: MultiLocation

TransferReserveAsset

Withdraw asset(s) (assets) from the ownership of origin and place equivalent assets under the ownership of dest within this consensus system (i.e. its sovereign account).

Send an onward XCM message to dest of ReserveAssetDeposited with the given effects.

assets: The asset(s) to be withdrawn.

dest: The location whose sovereign account will own the assets and thus the effective beneficiary for the assets and the notification target for the reserve asset deposit message.

effects: The orders that should be contained in the ReserveAssetDeposited which is sent onwards to dest.

Safety: No concerns.

Kind: Instruction.

Errors:

Fields of TransferReserveAsset

assets: MultiAssetsdest: MultiLocationeffects: Vec<Order<()>>

Transact

Apply the encoded transaction call, whose dispatch-origin should be origin as expressed by the kind of origin origin_type.

origin_type: The means of expressing the message origin as a dispatch origin.

max_weight: The weight of call; this should be at least the chain’s calculated weight and will be used in the weight determination arithmetic.

call: The encoded transaction to be applied.

Safety: No concerns.

Kind: Instruction.

Errors:

Fields of Transact

origin_type: OriginKindrequire_weight_at_most: u64call: DoubleEncoded<Call>

HrmpNewChannelOpenRequest

A message to notify about a new incoming HRMP channel. This message is meant to be sent by the relay-chain to a para.

sender: The sender in the to-be opened channel. Also, the initiator of the channel opening.

max_message_size: The maximum size of a message proposed by the sender.

max_capacity: The maximum number of messages that can be queued in the channel.

Safety: The message should originate directly from the relay-chain.

Kind: System Notification

Fields of HrmpNewChannelOpenRequest

sender: u32max_message_size: u32max_capacity: u32

HrmpChannelAccepted

A message to notify about that a previously sent open channel request has been accepted by the recipient. That means that the channel will be opened during the next relay-chain session change. This message is meant to be sent by the relay-chain to a para.

Safety: The message should originate directly from the relay-chain.

Kind: System Notification

Errors:

Fields of HrmpChannelAccepted

recipient: u32

HrmpChannelClosing

A message to notify that the other party in an open channel decided to close it. In particular, initiator is going to close the channel opened from sender to the recipient. The close will be enacted at the next relay-chain session change. This message is meant to be sent by the relay-chain to a para.

Safety: The message should originate directly from the relay-chain.

Kind: System Notification

Errors:

Fields of HrmpChannelClosing

initiator: u32sender: u32recipient: u32

RelayedFrom

A message to indicate that the embedded XCM is actually arriving on behalf of some consensus location within the origin.

Kind: Instruction

Errors:

Fields of RelayedFrom

who: InteriorMultiLocationmessage: Box<Xcm<Call>>

SubscribeVersion

Ask the destination system to respond with the most recent version of XCM that they support in a QueryResponse instruction. Any changes to this should also elicit similar responses when they happen.

Kind: Instruction

Fields of SubscribeVersion

query_id: u64max_response_weight: u64

UnsubscribeVersion

Cancel the effect of a previous SubscribeVersion instruction.

Kind: Instruction

Implementations

impl<Call> Xcm<Call>

pub fn into<C>(self) -> Xcm<C>

pub fn from<C>(xcm: Xcm<C>) -> Self