Module xcm::v2

Version 1 of the Cross-Consensus Message format data structures.

Re-exports

pub use super::v1::OriginKind;

Modules

opaque

prelude
A prelude for importing all types typically used when interacting with XCM messages.

Structs

Ancestor
A unit struct which can be converted into a MultiLocation of the inner parents value.

AncestorThen
A unit struct which can be converted into a MultiLocation of the inner parents value and the inner interior.

MultiAsset

MultiAssets
A Vec of MultiAssets. There may be no duplicate fungible items in here and when decoding, they must be sorted.

MultiLocation
A relative path between state-bearing consensus systems.

Parent
A unit struct which can be converted into a MultiLocation of parents value 1.

ParentThen
A tuple struct which can be converted into a MultiLocation of parents value 1 with the inner interior.

Xcm

Enums

AssetId
Classification of an asset being concrete or abstract.

AssetInstance
A general identifier for an instance of a non-fungible asset class.

BodyId
An identifier of a pluralistic body.

BodyPart
A part of a pluralistic body.

Error

Fungibility
Classification of whether an asset is fungible or not, along with a mandatory amount or instance.

Instruction
Cross-Consensus Message: A message from one consensus system to another.

Junction
A single item in a path to describe the relative location of a consensus system.

Junctions
Non-parent junctions that can be constructed, up to the length of 8. This specific Junctions implementation uses a Rust enum in order to make pattern matching easier.

MultiAssetFilter
MultiAsset collection, either MultiAssets or a single wildcard.

NetworkId
A global identifier of an account-bearing consensus system.

Outcome
Outcome of an XCM execution.

Response
Response data to a query.

SendError
Error result value when attempting to send an XCM message.

WeightLimit
An optional weight limit.

WildFungibility
Classification of whether an asset is fungible or not.

WildMultiAsset
A wildcard representing a set of assets.

Constants

VERSION
This module’s XCM version.

Traits

ExecuteXcm
Type of XCM message executor.

SendXcm
Utility for sending an XCM message.

XcmWeightInfo
The info needed to weight an XCM.

Type Definitions

InteriorMultiLocation
A relative location which is constrained to be an interior location of the context.

QueryId
An identifier for a query.

Result

SendResult
Result value when attempting to send an XCM message.

Weight
Local weight type; execution time in picoseconds.