Enum xcm::v2::Instruction

pub enum Instruction<RuntimeCall> {

    WithdrawAsset(MultiAssets),
    ReserveAssetDeposited(MultiAssets),
    ReceiveTeleportedAsset(MultiAssets),
    QueryResponse {
        query_id: QueryId,
        response: Response,
        max_weight: u64,
    },
    TransferAsset {
        assets: MultiAssets,
        beneficiary: MultiLocation,
    },
    TransferReserveAsset {
        assets: MultiAssets,
        dest: MultiLocation,
        xcm: Xcm<()>,
    },
    Transact {
        origin_type: OriginKind,
        require_weight_at_most: u64,
        call: DoubleEncoded<RuntimeCall>,
    },
    HrmpNewChannelOpenRequest {
        sender: u32,
        max_message_size: u32,
        max_capacity: u32,
    },
    HrmpChannelAccepted {
        recipient: u32,
    },
    HrmpChannelClosing {
        initiator: u32,
        sender: u32,
        recipient: u32,
    },
    ClearOrigin,
    DescendOrigin(InteriorMultiLocation),
    ReportError {
        query_id: QueryId,
        dest: MultiLocation,
        max_response_weight: u64,
    },
    DepositAsset {
        assets: MultiAssetFilter,
        max_assets: u32,
        beneficiary: MultiLocation,
    },
    DepositReserveAsset {
        assets: MultiAssetFilter,
        max_assets: u32,
        dest: MultiLocation,
        xcm: Xcm<()>,
    },
    ExchangeAsset {
        give: MultiAssetFilter,
        receive: MultiAssets,
    },
    InitiateReserveWithdraw {
        assets: MultiAssetFilter,
        reserve: MultiLocation,
        xcm: Xcm<()>,
    },
    InitiateTeleport {
        assets: MultiAssetFilter,
        dest: MultiLocation,
        xcm: Xcm<()>,
    },
    QueryHolding {
        query_id: QueryId,
        dest: MultiLocation,
        assets: MultiAssetFilter,
        max_response_weight: u64,
    },
    BuyExecution {
        fees: MultiAsset,
        weight_limit: WeightLimit,
    },
    RefundSurplus,
    SetErrorHandler(Xcm<RuntimeCall>),
    SetAppendix(Xcm<RuntimeCall>),
    ClearError,
    ClaimAsset {
        assets: MultiAssets,
        ticket: MultiLocation,
    },
    Trap(u64),
    SubscribeVersion {
        query_id: QueryId,
        max_response_weight: u64,
    },
    UnsubscribeVersion,

}

Cross-Consensus Message: A message from one consensus system to another.

Consensus systems that may send and receive messages include blockchains and smart contracts.

All messages are delivered from a known origin, expressed as a MultiLocation.

This is the inner XCM format and is version-sensitive. Messages are typically passed using the outer XCM format, known as VersionedXcm.

Variants

WithdrawAsset(MultiAssets)

Withdraw asset(s) (assets) from the ownership of origin and place them into the Holding Register.

assets: The asset(s) to be withdrawn into holding.

Kind: Command.

Errors:

ReserveAssetDeposited(MultiAssets)

Asset(s) (assets) have been received into the ownership of this system on the origin system and equivalent derivatives should be placed into the Holding Register.

assets: The asset(s) that are minted into holding.

Safety: origin must be trusted to have received and be storing assets such that they may later be withdrawn should this system send a corresponding message.

Kind: Trusted Indication.

Errors:

ReceiveTeleportedAsset(MultiAssets)

Asset(s) (assets) have been destroyed on the origin system and equivalent assets should be created and placed into the Holding Register.

assets: The asset(s) that are minted into the Holding Register.

Safety: origin must be trusted to have irrevocably destroyed the corresponding assets prior as a consequence of sending this message.

Kind: Trusted Indication.

Errors:

QueryResponse

Fields

query_id: QueryId

response: Response

max_weight: u64

Respond with information that the local system is expecting.

query_id: The identifier of the query that resulted in this message being sent.

response: The message content.

max_weight: The maximum weight that handling this response should take.

Safety: No concerns.

Kind: Information.

Errors:

TransferAsset

Fields

assets: MultiAssets

beneficiary: MultiLocation

Withdraw asset(s) (assets) from the ownership of origin and place equivalent assets under the ownership of beneficiary.

assets: The asset(s) to be withdrawn.

beneficiary: The new owner for the assets.

Safety: No concerns.

Kind: Command.

Errors:

TransferReserveAsset

Fields

assets: MultiAssets

dest: MultiLocation

xcm: Xcm<()>

Withdraw asset(s) (assets) from the ownership of origin and place equivalent assets under the ownership of dest within this consensus system (i.e. its sovereign account).

Send an onward XCM message to dest of ReserveAssetDeposited with the given xcm.

assets: The asset(s) to be withdrawn.

dest: The location whose sovereign account will own the assets and thus the effective beneficiary for the assets and the notification target for the reserve asset deposit message.

xcm: The instructions that should follow the ReserveAssetDeposited instruction, which is sent onwards to dest.

Safety: No concerns.

Kind: Command.

Errors:

Transact

Fields

origin_type: OriginKind

require_weight_at_most: u64

call: DoubleEncoded<RuntimeCall>

Apply the encoded transaction call, whose dispatch-origin should be origin as expressed by the kind of origin origin_type.

origin_type: The means of expressing the message origin as a dispatch origin.

max_weight: The weight of call; this should be at least the chain’s calculated weight and will be used in the weight determination arithmetic.

call: The encoded transaction to be applied.

Safety: No concerns.

Kind: Command.

Errors:

HrmpNewChannelOpenRequest

Fields

sender: u32

max_message_size: u32

max_capacity: u32

A message to notify about a new incoming HRMP channel. This message is meant to be sent by the relay-chain to a para.

sender: The sender in the to-be opened channel. Also, the initiator of the channel opening.

max_message_size: The maximum size of a message proposed by the sender.

max_capacity: The maximum number of messages that can be queued in the channel.

Safety: The message should originate directly from the relay-chain.

Kind: System Notification

HrmpChannelAccepted

Fields

recipient: u32

A message to notify about that a previously sent open channel request has been accepted by the recipient. That means that the channel will be opened during the next relay-chain session change. This message is meant to be sent by the relay-chain to a para.

Safety: The message should originate directly from the relay-chain.

Kind: System Notification

Errors:

HrmpChannelClosing

Fields

initiator: u32

sender: u32

recipient: u32

A message to notify that the other party in an open channel decided to close it. In particular, initiator is going to close the channel opened from sender to the recipient. The close will be enacted at the next relay-chain session change. This message is meant to be sent by the relay-chain to a para.

Safety: The message should originate directly from the relay-chain.

Kind: System Notification

Errors:

ClearOrigin

Clear the origin.

This may be used by the XCM author to ensure that later instructions cannot command the authority of the origin (e.g. if they are being relayed from an untrusted source, as often the case with ReserveAssetDeposited).

Safety: No concerns.

Kind: Command.

Errors:

DescendOrigin(InteriorMultiLocation)

Mutate the origin to some interior location.

Kind: Command

Errors:

ReportError

Fields

query_id: QueryId

dest: MultiLocation

max_response_weight: u64

Immediately report the contents of the Error Register to the given destination via XCM.

A QueryResponse message of type ExecutionOutcome is sent to dest with the given query_id and the outcome of the XCM.

query_id: An identifier that will be replicated into the returned XCM message.

dest: A valid destination for the returned XCM message.

max_response_weight: The maximum amount of weight that the QueryResponse item which is sent as a reply may take to execute. NOTE: If this is unexpectedly large then the response may not execute at all.

Kind: Command

Errors:

DepositAsset

Fields

assets: MultiAssetFilter

max_assets: u32

beneficiary: MultiLocation

Remove the asset(s) (assets) from the Holding Register and place equivalent assets under the ownership of beneficiary within this consensus system.

assets: The asset(s) to remove from holding.

max_assets: The maximum number of unique assets/asset instances to remove from holding. Only the first max_assets assets/instances of those matched by assets will be removed, prioritized under standard asset ordering. Any others will remain in holding.

beneficiary: The new owner for the assets.

Kind: Command

Errors:

DepositReserveAsset

Fields

assets: MultiAssetFilter

max_assets: u32

dest: MultiLocation

xcm: Xcm<()>

Remove the asset(s) (assets) from the Holding Register and place equivalent assets under the ownership of dest within this consensus system (i.e. deposit them into its sovereign account).

Send an onward XCM message to dest of ReserveAssetDeposited with the given effects.

assets: The asset(s) to remove from holding.

max_assets: The maximum number of unique assets/asset instances to remove from holding. Only the first max_assets assets/instances of those matched by assets will be removed, prioritized under standard asset ordering. Any others will remain in holding.

dest: The location whose sovereign account will own the assets and thus the effective beneficiary for the assets and the notification target for the reserve asset deposit message.

xcm: The orders that should follow the ReserveAssetDeposited instruction which is sent onwards to dest.

Kind: Command

Errors:

ExchangeAsset

Fields

give: MultiAssetFilter

receive: MultiAssets

Remove the asset(s) (give) from the Holding Register and replace them with alternative assets.

The minimum amount of assets to be received into the Holding Register for the order not to fail may be stated.

give: The asset(s) to remove from holding.

receive: The minimum amount of assets(s) which give should be exchanged for.

Kind: Command

Errors:

InitiateReserveWithdraw

Fields

assets: MultiAssetFilter

reserve: MultiLocation

xcm: Xcm<()>

Remove the asset(s) (assets) from holding and send a WithdrawAsset XCM message to a reserve location.

assets: The asset(s) to remove from holding.

reserve: A valid location that acts as a reserve for all asset(s) in assets. The sovereign account of this consensus system on the reserve location will have appropriate assets withdrawn and effects will be executed on them. There will typically be only one valid location on any given asset/chain combination.

xcm: The instructions to execute on the assets once withdrawn on the reserve location.

Kind: Command

Errors:

InitiateTeleport

Fields

assets: MultiAssetFilter

dest: MultiLocation

xcm: Xcm<()>

Remove the asset(s) (assets) from holding and send a ReceiveTeleportedAsset XCM message to a dest location.

assets: The asset(s) to remove from holding.

dest: A valid location that respects teleports coming from this location.

xcm: The instructions to execute on the assets once arrived on the destination location.

NOTE: The dest location MUST respect this origin as a valid teleportation origin for all assets. If it does not, then the assets may be lost.

Kind: Command

Errors:

QueryHolding

Fields

query_id: QueryId

dest: MultiLocation

assets: MultiAssetFilter

max_response_weight: u64

Send a Balances XCM message with the assets value equal to the holding contents, or a portion thereof.

query_id: An identifier that will be replicated into the returned XCM message.

dest: A valid destination for the returned XCM message. This may be limited to the current origin.

assets: A filter for the assets that should be reported back. The assets reported back will be, asset-wise, the lesser of this value and the holding register. No wildcards will be used when reporting assets back.

max_response_weight: The maximum amount of weight that the QueryResponse item which is sent as a reply may take to execute. NOTE: If this is unexpectedly large then the response may not execute at all.

Kind: Command

Errors:

BuyExecution

Fields

fees: MultiAsset

weight_limit: WeightLimit

Pay for the execution of some XCM xcm and orders with up to weight picoseconds of execution time, paying for this with up to fees from the Holding Register.

fees: The asset(s) to remove from the Holding Register to pay for fees.

weight_limit: The maximum amount of weight to purchase; this must be at least the expected maximum weight of the total XCM to be executed for the AllowTopLevelPaidExecutionFrom barrier to allow the XCM be executed.

Kind: Command

Errors:

RefundSurplus

Refund any surplus weight previously bought with BuyExecution.

Kind: Command

Errors: None.

SetErrorHandler(Xcm<RuntimeCall>)

Set the Error Handler Register. This is code that should be called in the case of an error happening.

An error occurring within execution of this code will NOT result in the error register being set, nor will an error handler be called due to it. The error handler and appendix may each still be set.

The apparent weight of this instruction is inclusive of the inner Xcm; the executing weight however includes only the difference between the previous handler and the new handler, which can reasonably be negative, which would result in a surplus.

Kind: Command

Errors: None.

SetAppendix(Xcm<RuntimeCall>)

Set the Appendix Register. This is code that should be called after code execution (including the error handler if any) is finished. This will be called regardless of whether an error occurred.

Any error occurring due to execution of this code will result in the error register being set, and the error handler (if set) firing.

The apparent weight of this instruction is inclusive of the inner Xcm; the executing weight however includes only the difference between the previous appendix and the new appendix, which can reasonably be negative, which would result in a surplus.

Kind: Command

Errors: None.

ClearError

Clear the Error Register.

Kind: Command

Errors: None.

ClaimAsset

Fields

assets: MultiAssets

ticket: MultiLocation

Create some assets which are being held on behalf of the origin.

assets: The assets which are to be claimed. This must match exactly with the assets claimable by the origin of the ticket.

ticket: The ticket of the asset; this is an abstract identifier to help locate the asset.

Kind: Command

Errors:

Trap(u64)

Always throws an error of type Trap.

Kind: Command

Errors:

Trap: All circumstances, whose inner value is the same as this item’s inner value.

SubscribeVersion

Fields

query_id: QueryId

max_response_weight: u64

Ask the destination system to respond with the most recent version of XCM that they support in a QueryResponse instruction. Any changes to this should also elicit similar responses when they happen.

query_id: An identifier that will be replicated into the returned XCM message.

max_response_weight: The maximum amount of weight that the QueryResponse item which is sent as a reply may take to execute. NOTE: If this is unexpectedly large then the response may not execute at all.

Kind: Command

Errors: Fallible

UnsubscribeVersion

Cancel the effect of a previous SubscribeVersion instruction.

Kind: Command

Errors: Fallible

Implementations

impl<RuntimeCall> Instruction<RuntimeCall>

pub fn into<C>(self) -> Instruction<C>

pub fn from<C>(xcm: Instruction<C>) -> Self